pass in entire dataset object to ConceptSetCreator, move mocked data to test constants
