+++ src/dataset-builder/ConceptSelector.ts
@@ -6,8 +6,9 @@ import { ActionBar } from 'src/components/ActionBar';
 import { Link } from 'src/components/common';
 import { icon } from 'src/components/icons';
 import { TreeGrid } from 'src/components/TreeGrid';
+import { getMockConceptForId } from 'src/dataset-builder/TestConstants';
 import { SnapshotBuilderConcept as Concept } from 'src/libs/ajax/DataRepo';
-import { DatasetBuilder, getConceptForId } from 'src/libs/ajax/DatasetBuilder';
+import { DatasetBuilder } from 'src/libs/ajax/DatasetBuilder';
 
 import { PAGE_PADDING_HEIGHT, PAGE_PADDING_WIDTH } from './constants';
 
@@ -73,7 +74,7 @@ export const ConceptSelector = (props: ConceptSelectorProps) => {
       h(ActionBar, {
         prompt: cart.length === 1 ? '1 concept selected' : `${cart.length} concepts selected`,
         actionText,
-        onClick: () => _.flow(_.map(getConceptForId), onCommit)(cart),
+        onClick: () => _.flow(_.map(getMockConceptForId), onCommit)(cart),
       }),
   ]);
 };
